fixed chapter 11 q20 and 21
previously we thought since stopwords changes in 2015 that caused problem in course quiz results but it was not so. now we are using the old file only and just changed code to remove bugs and we are getting correct results. now from here we will go do chpater 12 quizes again from 0.